refactor: Handler dedup, error helpers, award period, and test infra (#22, #24, #31, #7)
- #22: Extract apply_common_filters() in activation.rs to reduce duplication
- #24: Add db_error(), tx_error(), row_not_found() helper functions in error.rs
- #31: Replace hardcoded dates with AwardPeriod struct (2025/6/1-12/31 JST)
- #7: Add 21 unit tests (common/utils.rs: 12, service/model/award.rs: 9)
- Extract award_calculator module from user_service.rs
