+++ api/src/handler/activation.rs
@@ -24,45 +24,54 @@ const CACHE_TTL_SPOTS: i64 = 30;
 const CACHE_TTL_ALERTS: i64 = 180;
 const CACHE_TTL_TRACK: i64 = 60;
 
+/// パラメータからFindActBuilderにグルーピングとフィルタを適用
+fn apply_common_filters(param: &GetParam, mut query: FindActBuilder, default_hours: i64) -> FindActBuilder {
+    // グルーピング設定
+    if let Some(callsign) = &param.by_call {
+        if callsign.starts_with("null") {
+            query = query.group_by_callsign(None)
+        } else {
+            query = query.group_by_callsign(Some(callsign.clone()))
+        }
+    } else if let Some(reference) = &param.by_ref {
+        if reference.starts_with("null") {
+            query = query.group_by_reference(None)
+        } else {
+            query = query.group_by_reference(Some(reference.clone()))
+        }
+    } else {
+        query = query.group_by_callsign(None)
+    }
+
+    // 時間フィルタ
+    let hours = param.hours_ago.unwrap_or(default_hours);
+    query = query.issued_after(Utc::now() - Duration::hours(hours));
+
+    // パターンフィルタ
+    if let Some(pat) = &param.pat_ref {
+        query = query.pattern(pat);
+    }
+
+    // ログIDフィルタ
+    if let Some(log_id) = &param.log_id {
+        query = query.log_id(log_id);
+    }
+
+    query
+}
+
 async fn show_spots(
     user_service: Inject<AppRegistry, dyn UserService>,
     kvs_repo: Inject<AppRegistry, dyn KvsRepositry>,
     param: GetParam,
-    mut query: FindActBuilder,
+    query: FindActBuilder,
 ) -> AppResult<Json<Value>> {
     let key = param.to_key();
     if let Some(val) = kvs_repo.get(&key).await {
         return Ok(Json(val));
     };
 
-    if let Some(callsign) = param.by_call {
-        if callsign.starts_with("null") {
-            query = query.group_by_callsign(None)
-        } else {
-            query = query.group_by_callsign(Some(callsign))
-        }
-    } else if let Some(reference) = param.by_ref {
-        if reference.starts_with("null") {
-            query = query.group_by_reference(None)
-        } else {
-            query = query.group_by_reference(Some(reference))
-        }
-    } else {
-        query = query.group_by_callsign(None)
-    }
-
-    let hours = param.hours_ago.unwrap_or(3);
-    query = query.issued_after(Utc::now() - Duration::hours(hours));
-
-    if let Some(pat) = param.pat_ref {
-        query = query.pattern(&pat);
-    }
-
-    if let Some(log_id) = param.log_id {
-        query = query.log_id(&log_id);
-    }
-
-    let query = query.build();
+    let query = apply_common_filters(&param, query, 3).build();
 
     let result = user_service.find_spots(query).await?;
     let mut spots: Vec<_> = result
@@ -87,37 +96,14 @@ async fn show_alerts(
     user_service: Inject<AppRegistry, dyn UserService>,
     kvs_repo: Inject<AppRegistry, dyn KvsRepositry>,
     param: GetParam,
-    mut query: FindActBuilder,
+    query: FindActBuilder,
 ) -> AppResult<Json<Value>> {
     let key = param.to_key();
     if let Some(val) = kvs_repo.get(&key).await {
         return Ok(Json(val));
     };
 
-    if let Some(callsign) = param.by_call {
-        if callsign.starts_with("null") {
-            query = query.group_by_callsign(None)
-        } else {
-            query = query.group_by_callsign(Some(callsign))
-        }
-    } else if let Some(reference) = param.by_ref {
-        if reference.starts_with("null") {
-            query = query.group_by_reference(None)
-        } else {
-            query = query.group_by_reference(Some(reference))
-        }
-    } else {
-        query = query.group_by_callsign(None)
-    }
-
-    if let Some(pat) = param.pat_ref {
-        query = query.pattern(&pat);
-    }
-
-    let hours = param.hours_ago.unwrap_or(24);
-    query = query.issued_after(Utc::now() - Duration::hours(hours));
-
-    let query = query.build();
+    let query = apply_common_filters(&param, query, 24).build();
 
     let result = user_service.find_alerts(query).await?;
     let mut alerts: Vec<_> = result
