+++ service/src/implement/mod.rs
@@ -1,4 +1,5 @@
 pub mod admin_periodic;
 pub mod admin_service;
 pub mod aprs_service;
+pub mod award_calculator;
 pub mod user_service;
